src/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

# -i https://pypi.org/simple
aiofiles==0.6.0
aiosqlite==0.19.0
alembic==1.10.3
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
anyio==3.6.2; python_full_version >= '3.6.2'
argparse==1.4.0
async-generator==1.10; python_version >= '3.5'
attrs==23.1.0; python_version >= '3.7'
authlib==0.15.6
auto-archiver==0.5.6
beautifulsoup4==4.12.2; python_version >= '3.6'
billiard==3.6.4.0
boto3==1.26.115; python_version >= '3.7'
botocore==1.29.115; python_version >= '3.7'
brotli==1.0.9; platform_python_implementation == 'CPython'
bs4==0.0.1
cachetools==5.3.0; python_version ~= '3.7'
celery==4.4.7
certifi==2022.12.7; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3; python_version >= '3.7'
cloudscraper==1.2.69
cryptography==38.0.4; python_version >= '3.6'
dataclasses-json==0.5.7; python_version >= '3.6'
dateparser==1.1.8; python_version >= '3.7'
exceptiongroup==1.1.1; python_version < '3.11'
fastapi-utils==0.2.1
fastapi==0.95.1
ffmpeg-python==0.2.0
filelock==3.12.0; python_version >= '3.7'
flask==2.2.3; python_version >= '3.7'
flower==0.9.7
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core==2.11.0; python_version >= '3.7'
google-api-python-client==2.86.0; python_version >= '3.7'
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0; python_version >= '3.6'
google-auth==2.17.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.59.0; python_version >= '3.7'
greenlet==2.0.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gspread==5.8.0; python_version not in '3.0, 3.1, 3.2, 3.3' and python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
humanize==4.6.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
instaloader==4.9.6; python_version >= '3.8'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2
jmespath==1.0.1; python_version >= '3.7'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
loguru==0.7.0
lxml==4.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.2.4; python_version >= '3.7'
markdown-it-py==2.2.0; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
marshmallow-enum==1.5.1
marshmallow==3.19.0; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
mutagen==1.46.0; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
oauth2client==4.1.3
oauthlib==3.2.2; python_version >= '3.6'
outcome==1.2.0; python_version >= '3.7'
packaging==23.1; python_version >= '3.7'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prometheus-client==0.8.0
protobuf==4.22.3; python_version >= '3.7'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyaes==1.6.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.17; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.10.7; python_version >= '3.7'
pygments==2.15.0; python_version >= '3.7'
pyparsing==3.0.9; python_version >= '3.1'
pysocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-slugify==8.0.1; python_version >= '3.7'
python-twitter-v2==0.8.1; python_version >= '3.6' and python_version < '4'
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2023.3
pyyaml==6.0; python_version >= '3.6'
redis==3.5.3
regex==2023.3.23; python_version >= '3.8'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests-toolbelt==0.10.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.2
rich==13.3.4; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
s3transfer==0.6.0; python_version >= '3.7'
selenium==4.8.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
snscrape==0.6.2.20230320; python_version ~= '3.8'
sortedcontainers==2.4.0
soupsieve==2.4.1; python_version >= '3.7'
sqlalchemy==1.4.47
starlette==0.26.1; python_version >= '3.7'
telethon==1.28.5; python_version >= '3.5'
text-unidecode==1.3
tiktok-downloader==0.3.4
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.3; python_full_version >= '3.5.2'
tqdm==4.65.0; python_version >= '3.7'
trio-websocket==0.10.2; python_version >= '3.7'
trio==0.22.0; python_version >= '3.7'
typing-extensions==4.5.0; python_version >= '3.7'
typing-inspect==0.8.0
tzdata==2023.3; python_version >= '3.6'
tzlocal==4.3; python_version >= '3.7'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.21.1
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
vk-api==11.9.9
vk-url-scraper==0.3.15; python_version >= '3.7'
websockets==10.4; python_version >= '3.7'
werkzeug==2.2.3; python_version >= '3.7'
wsproto==1.2.0; python_version >= '3.7'
yt-dlp==2023.2.17; python_version >= '3.7'
